static casting - fix erro

+++ targets/simple_switch/simple_switch.cpp
@@ -80,7 +80,7 @@ uint32_t get_probability_sum(const char *buf, size_t s) {
 uint32_t get_probability_path(const char *buf, size_t s, uint32_t random) {
   uint32_t default_route = 0;
   for (size_t i = 0; i < s; i++) {
-    if (random < (0 | buf[i])) {
+    if (random < static_cast<uint32_t>(0 | buf[i])) {
       default_route = static_cast<uint32_t>(i);
       break;
     }
